chatbot/requirements.txt
langchain
langchain-openai
langgraph
curl_cffi
newspaper3k
python-dotenv
